+++ src/main/java/com/almostreliable/kubeoritech/event/particle/state/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package com.almostreliable.kubeoritech.event.particle.state;
+
+import org.lwjgl.system.NonnullDefault;